full tenant CRUD routes

+++ src/user/user.service.ts
@@ -4,6 +4,9 @@ import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { Prisma } from '@prisma/client';
 
+
+// TODO: Add password hashing and validation logic
+
 @Injectable()
 export class UserService {
   constructor(private readonly prisma: PrismaService) {}
@@ -17,13 +20,12 @@ export class UserService {
 
     if (isUserExists) throw new ConflictException('User already exists'); // returns 409 Conflict
 
-    // TODO: Hash password here
-    
     const dbUser = {
       email: createUserDto.email,
       password_hash: createUserDto.password, // This should be hashed before saving
       full_name: createUserDto.full_name,
       role: createUserDto.role,
+      tenant_id: createUserDto.tenant_id, // Optional, if user is created within a tenant context
     }
 
     return this.prisma.user.create({
@@ -33,6 +35,7 @@ export class UserService {
         email: true,
         full_name: true,
         role: true,
+        tenant_id: true, // Optional, if user is created within a tenant context
       },
     });
   }
@@ -66,6 +69,20 @@ export class UserService {
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+
+    // Check if at least one field is provided for update
+    if (!updateUserDto.email && !updateUserDto.full_name && !updateUserDto.role && !updateUserDto.tenant_id) {
+      throw new ConflictException('No fields to update'); // returns 409 Conflict
+    }
+
+    // Check if user exists
+    const foundUser = await this.prisma.user.findUnique({
+      where: { id },
+      });
+    
+    if (!foundUser) throw new NotFoundException('User not found'); // returns 404 Not Found
+
+    // Update user with provided fields
     const updatedUser = await this.prisma.user.update({
       where: { id },
       data: updateUserDto,
@@ -74,6 +91,7 @@ export class UserService {
         email: true,
         full_name: true,
         role: true,
+        tenant_id: true,
       },
     });
 
@@ -86,6 +104,14 @@ export class UserService {
   }
 
   async remove(id: string) {
+    // Check if user exists
+    const foundUser = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!foundUser) throw new NotFoundException('User not found'); // returns 404 Not Found
+
+    // Delete user
     const deletedUser = await this.prisma.user.delete({
       where: { id },
       select: {
